requirements.txt
azure-functions
azure-cosmos==4.2.0